Initial an application with a moving ball in a triangular area
Add App module, to create an application with a moving ball in a triangular area.
Add method start() to App module, to start the ball rolling on the canvas.
Add method stop() to App module, to reset the variable values to the original values.
Add method validate() to App module, to check the validity of the entered data in the widget.
Add method save_old_value() to App module, to save the previous widget value.
Add method modal_help() to App module, to display a modal window with help text.
Add method modal_about() to App module, to display the "About" modal window.
Add MovingBall module, representing a moving ball on a canvas.
Add method distance_to_segment() to MovingBall module, to calculates the distance from a ball to an edge on the canvas.
Add method move_bal()l to MovingBall module, to move the ball across the canvas and reflect it off the walls if it collides with them.
Add method clear_ball() to MovingBall module, to remove the balloon and all other objects on the canvas.
Add Validator module, to verify the data entered.
Add method validate_int() to Validator module, to check if the value is an integer.
Add Controller module.
Add method get_version() to Controller module, to return the system version.
Add method get_help() to Controller module, for the return of the programme certificate.
Add Storage module, representing a repository.
Add method read_version() to Storage module, to read a version from a file.
Add method read_help() to Storage module, to read help from the file.
Add version.txt, to store a version.
Add help.txt, to store a help.
Add gif demonstrating the work of the programme.